wlan: libra: Fix for hidden SSID issue
Fix checking for bogus SSID logic. 0x20 <blank> is
the smallest displayable ASCII hex value and 0x7E
is the largest displayable ASCII value.  If the
character is not within the displayable ASCII range
of values, then the SSID is bogus. Also consider
ASCII 0 a valid character to allow connection to
Cerberus.