BT/WLAN Coex: Tell coex when WLAN is going to go down
The WLAN driver needs to send a netlink message telling BTC it is
going down, so that BTC will close the netlink socket, releasing
the use count on the WLAN driver module, so that it can be unloaded.
A new ioctl is used to tell the driver to send the netlink message.
Part of the solution for CR 240174.

Change-Id: I3eafd822c3fe1709e0fc08d1b3d7d4322d835217