Voting for PWM mode of PMAPP_VREG_S4 for 7x30.
Voting for PWM mode of PMAPP_VREG_S4 for 7x30.
While powering up the libra we need to vote for
PWM mode of PMAPP_VREG_S4, as due to optimization
of PMIC this may be by default in PFM mode.
While putting wlan driver in deep sleep, we are
voting off forPWM mode (Don't care) of PMAPP_VREG_S4.

Change-Id: Ibe00dc5c4e96037e6013f9c69a1c9daa7776655f