wlan: Create the M= dir to avoid unnecessary recompilation
The wlan component unnecessarily gets rebuilt the second time
(without any source code changes) because of change in the
command line used to build the wlan sources. This happens
because kbuild attemps to create .tmp files before creating
the directory in which they are supposed to reside. Creating
the M= directory in the wlan Android.mk and making each module
depend on that target solves this problem

Targets tested on: msm7630_fusion

Tests:
1) Built wlan modules with 'make <module_target> from android
root twice and verified it does not unnecessarily recompile
the second time
2) Verified a full android build

Change-Id: I4af825f07df9506f3f0f3c6ed41f367bde66ab81