wlan: Code Complete Release of the 8960 WLAN Driver
This is the "Code Complete" release of the msm8960 WLAN Driver for
the benefit of sharing with ODM partners.  Although the underlying
code is functional, not all of the infrastructure is in place to
support unassisted initialization of the driver (i.e. PIL support
for loading the RIVA software image) and hence this should be
considered a "code-only" release.

Change-Id: Ic2cda85f13168e919a0f807410022ac994b231ed