Build utils from the utils directory
Modify the build to pick up the utils from the utils directory
rather than from the board specific directories.  Modify the make
files to ensure that the ftm directory symlink is made since this
won't make it through the packing process.  This with an associated
change in the pack rules should mean HY11 has WLAN sources that
build

Change-Id: I07db139a4c670c6df6b6fc5f9a1b55667e4158bd