wlan: NULL pointer dereference happens when wifi in ON state.
The issue happens when STA is connected to AP with WPA security mode and just leave the phone.
Try to access internet every 5mins or more than that. Very rarely NULL pointer access happening at "hdd_RoamSetKeyCompleteHandler" and
this leads the kernel panic.

The issue might be happening beacuse of the deauth/disconnect from the AP after a successful connection,
where SME is in disconnected state where HDD have not been informed. This may happen when HDD sets the key,
after PE posts a deauth_ind to SME. Due to this csrIsSetKeyAllowed fails and calls csrRoamCallCallback function.
Finally the control reach hdd_RoamSetKeyCompleteHandler where the function try to access pRoamInfo->fAuthRequired.
At this particular scenario, pRoamInfo may be NULL which is causing the issue.

Change-Id: Id9bae116237602ecc08572efb485c44560b17acb
CRs-fixed: 292467