wlan: STA is not able to connect hidden SSID.
When we are in Android Home screen, Android GUI sends scan-passive command to the driver and sets the
scan mode to passive. if any ssid is configured in Add Wi-Fi network of GUI, then combo scan recieves
the ssid from supplicant.But as we are in home screen and scan mode is passive STA won't be able to
connect to hidden ssid.

To fix this issue, added a check in iw_set_cscan function for number of ssids in CSCAN buffer.
if the value is positive then set the scan mode to active. So driver sends unicast probe request
to hidden ssid AP and helps to connect to AP.

Change-Id: Idf173456ddb23f5c5d1e5a9eb9b60b48333cca27
CRs-fixed: 298780