wlan: Unconditionally set -Wframe-larger-than
The logic to add -Wframe-larger-than to the EXTRA_CFLAGS was
originally added as conditional code since, at the time it was added,
not all of the versions of Android were bundled with a compiler which
supported the option.  Now that all current versions of Android
support the option, make it unconditional.

CRs-Fixed: 318134
Change-Id: I51ea410d03ebb41d5b4ceac6742c3717cdc81a96