Add temporary code to handle new version of ClassA stats
In an upcoming version of RIVA the ClassA stats will contain two
additional members.  We add temporary code to handle those members so
that when that version of RIVA is deployed, the stock LA AU will be
able to handle them.  We will subsequently deploy a new version of the
host driver that will only expect the new ClassA format.

Change-Id: I952116d33eb28d29b7f4cd8b5ecfef238c3eb6a9