wlan: CS 1.0 preliminary release #18
Integrate the following CLs from Perforce:

609959: Fixes for certain P2P TX Frames
1) When concurrency is enabled, some action frames sent from
supplicant on GO interface are getting dropped at the hdd, when there
is no active session associated with that interface. But we should
allow these frames over the air in order to complete the connection
between p2p-client and GO.  Changes done to transmit the action frames
when at least one active session exists.
2) In GO mode supplicant is sending Go Discovery Request and waiting
for tx status from the driver. After successfully transmitting this
frame over air, driver is returning the status '0'(success) to
supplicant.  But supplicant is expecting success status as '1'.  This
had been taken care in STA mode, but in SAP/GO mode the status
conversion is not done before returning the tx status to supplicant
and this was leading supplicant to not send the device discovery
request over the air.  Changes done to convert this tx status before
returning the status to supplicant.

Change-Id: I7e8ad795d7e0a84eec3ccd1d5f42787cd33cda69
CRs-Fixed: 326523