wlan: CS 1.0 preliminary release #19
Integrate the following CLs from Perforce:

609923: BT-AMP PAL qualification fix
For TP/PL/PLC/BV-03-C Physical Link complete event states the wrong
Reason code for failure. Fixed it so that PAL returns the correct
timeout error code to HCI.

610151: BT-AMP PAL qualification fix
HCI_IO_TRANS_ERROR seen while sending the
HCI_READ_LOCATION_DATA_OPCODE command (implementation of read/write
location data).  Fixed it so that PAL returns correct response to HCI
for location data cmd

610213: Changes for MIC Failure and Counter Measures
As per WFA certification requirements, if multiple MIC failures occur
within a period of 60 seconds the authenticator should deauthenticate
all STAs. To achieve this, hostapd needs to be notified when a MIC
failure is detected. Also, the MIC counter measure indication from
hostapd needs to be handled by the WLAN driver. This change includes:
- Passing MIC failure indication to nl80211
- Support for handling ioctl - SIOCSIWAUTH

610294: Fix for PNO non-bcast SSID
Added ability to scan for non-bcast networks in PNO

610306: Integration of revised RIVA API

610467: Integration of revised RIVA API

610469: KW fix for ICS WAPI code

610488: Fix for Few STA's are taking long time to connect to SAP
Following are the reasons for STA's are taking long time to connect to SAP.
1) Probe responses are being sent through ACK BTQM 9, This result in
retrying the probe reponses and filling up BTQM because of STA's
scanned and moved out of the channel. This result in AUTH frames not
coming OTA even though the PE is responded with AUTH frame. To fix
this issue mapping the probe responses to NO ACK BTQM.
2) If Disassoc frame is missed from STA, SAP is not able to detect
that the STA is not there. Added the changes for link monitoring in
RIVA. After the changes added in RIVA still because of a issue in LIM,
eWNI_SME_DISASSOC_REQ message is getting dropeed. To fix this issue
added the code to populate doNotSendOverTheAir in
eWNI_SME_DISASSOC_REQ.

610543: Integration of revised RIVA API

610667: BT-AMP Pal Qualification fix
The default flow control mode for AMP should be set to 'Data block
based flow control'.  Following changes are done to achieve this:
1)Return same value for Max_PDU_Size in Read Local AMP Info as it
returns for Data Block Length in Read Data Block Size
2)Report block-based in the Read Flow Control Mode command
3)Return the Number of Completed Data Blocks event instead of the
Number of Completed Packets event. In this event the Num Completed
Blocks field and the Num Completed Packets fields will be equal, and
will be the same as what PAL has previously been returning in the Num
Completed Packets field of the Number of Completed Packets event. The
Total Num Data Blocks field should be set to the same value as
returned in the Total Num Data Blocks field of the Read Data Block
Size command. This value is always the same for every event.
Also one KW fix

610833: Not to do roaming for SAP/GO when its peer station disassociated
Fix random crashes cause by aborting scan when remain-on-channel is
being process by LIM

610834: Updates to P2P logic
1) Changes done to send action frame on other frequency in P2P GO
Case.  After starting Autonomous GO, when we try to invite peer
device, we are sending action frame on channel where GO is
beaconing. In this case peer device is on other channel. it is not
able to receive invitation action frame. Now While sending action
frame, if Go Operating channel and Peer channel is different then we
will do remain on channel on peer operating channel and send action so
it can receive the invitation request.
2) Set NOA for remain on channel time duration for sending invitation
action frame.

610835: Fix issue with ERP bit setting
Fixed "ERP bit is not set when the B STA is connected to SoftAP in G
mode".  Due to extra ht capability check, driver was not able to
enable the protection for B STA when SAP is configured in G mode.

610836: Moved the ht function calls under htcapability flag check

Change-Id: I2f7713226df17b0be6c94258ec17ee756fc73a0c
CRs-Fixed: 273033,274014,327470,329667,330309,330464,331094