wlan: CS 1.0 preliminary release #23
Integrate the following CLs from Perforce:

611861: Fix issue with SoftAP interface TX queue enable logic
A kernel panic was observed immediately following turning off WiFi
Hotspot mode.  The stack trace revealed a fault with the "station
deinit" logic, namely that there is logic which can enable the
interface TX queue when a "suspended" station is deleted, but this
logic should only be invoked if the Hotspot is active.  When the
HotSpot is inactive, the interface must remain disabled.

611913: Integration of revised RIVA API

611924: Correct the order in which certain driver shutdown operations occur
When the driver is shutting down it performs the following operations:
   hdd_stop_all_adapters( pHddCtx );
   vos_stop( pVosContext );
   vos_sched_close( pVosContext );
   vos_close(pVosContext);
   hdd_close_all_adapters( pHddCtx );
One issue that was observed is that some of the activities that were
taking place during hdd_close_all_adapters() needed to be performed
prior to invoking vos_close().  So these activities were relocated and
are now invoked during hdd_stop_all_adapters().
In addition some debug code was added to assist in debugging future
driver load and unload issues.

612090: Modify wpalOpen() to handle error status from wpalDeviceInit()
A kernel panic was observed in the WLAN driver.  In the kernel logs
prior to the panic it was observed that a I/O memory remap operation
failed.  The function that was unable to perform the remap,
wpalDeviceInit(), correctly returned an error code to the caller,
wpalOpen().  However this error code was ignored.  So wpalOpen() was
modified to process the error code, and to propagate the error if it
occurs.

612130: Changes done while adding BSS entry during connect
Add BSS entry during connect if BSS entry is not present. After
successful BSS entry added and cfg80211 is informed with connect
result call cfg80211_put_bss so that memory can be freed.

612131: Fixed memory leaks
Fixed below memory leaks
1. In WLANTL_Open pTLCb->vosDummyBuf is allocated memory and this is
not freed in WLANTL_Close
2. cfg80211_put_bss is not being called after sending scan results of
each bss to cfg80211

612138: Don't check for SSID field for beacon frames during JOIN
In case of hidden SSID, beacons will be received without ssid
information, in those cases just BSSID check is sufficient. There is
no need to check for SSID match to proceed for join.

612139: Fix Assert in SAP stress test
Change code in SAP to set its state correctly to avoid state mismatch.
Remove assert and return error to avoid assertion in some tests.

612140: Removed hard tabs

612239: Pal Qualification fix
- When a Write Remote AMP assoc is issued with wrong TLV length, a
  Command Status event is to be sent to the host with Status 'Invalid
  HCI command parameters'.
- handle modify flow spec cmd
- handle enahnced flush cmd

612240: Changing the default value in the ini file to allow scanning
on DFS channels

612282: Remove "double free" of TL vosDummyBuf
CL 612131 integrated a change from the Volans codebase which frees the
TL vosDummyBuf in WLANTL_Close().  Unfortunately this buffer is
already being freed by WLANTL_CleanCB() (which is called by
WLANTL_Close()), resulting in a "double free" of the buffer which
leads to a list corruption.

612469: Fix race condition in WCTS_CloseTransport()
An exception was observed in the kernel log, with the stack trace
pointing to WCTS_CloseTransport().  This function has logic to
msleep() during rmmod to allow the SMD callback thread of execution to
finish executing within the wlan driver before the driver is unloaded.
Unfortunately the msleep() was initiated before the SMD callback had
occurred, so there was no guarantee that the callback would occur and
exit before the sleep time expired.  This is fixed so that we don't
msleep() until after the callback has occurred.  In addition, a recent
change requires that WCTS_CloseTransport() execute without serializing
work into the MC_Thread, so the WCTS module was modified to remove the
existing serialization of "close" processing.

612471: Fix memory leak in BTC processing
The WDA_McProcessMsg() drops an msg silently if the handler not
present. If there is a msg body for the un-handled msg it remains
un-freed. fixed it in the default case in the msg switch.

Change-Id: I6b62c07d1a0bc70e72e02eaf95333a4729f6d076
CRs-Fixed: 327539,330595,331009,331164,331165,331990,332747,333233,333837,334665,334872