wlan: CS 1.0 preliminary release #24
Integrate the following CLs from Perforce:
612294: A new API to clear all entries in all the ACL in case of SAP

612520: Conditionally enable debugging in the driver
Our driver supports two compile-time debug options, TIMER_MANAGER to
validate the proper use of timers, and MEMORY_DEBUG to validate the
proper use and reclaimation of dynamic memory.  This change will let
us know when the options have been enabled.

612527: Fixing compilation error due to incomplete checkin from CL612294

612541: potential Fix for BMU fatal error in SAP
It is observed that, in UAPSD mode when an AC has exhausted all its
weights, the dpuRF in TX BD is set according to trigger enabled flag
of next AC that should be serviced. This might result in TL setting
the DPU routing flag to firmware even if an AC is not trigger
enabled(but the next AC is trigger enabled). Changes done in TL to use
a temporary local variable to select the next AC that should be
serviced.

612553: Fixes for 802.11b testing
Fix the issue where short slot and short preamble were not getting
updated while doing addsta in softap mode. This was affecting 11b
testing.

612593: Implemented private ioctl for setting the frequency band
dynamically

612719: Not to allocate big size structure (above 1K) on the stack to
avoid stack overflow

Change-Id: I4172bd65f97c40f188d5b57829799b2bf29099e5
CRs-Fixed: 331700,333834,334262,334940,334949