wlan: CS 1.0 preliminary release #27
Integrate the following CLs from Perforce:

612519: Conditionally enable debugging in the driver
Our driver supports two compile-time debug options, TIMER_MANAGER to
validate the proper use of timers, and MEMORY_DEBUG to validate the
proper use and reclaimation of dynamic memory.  However these options
were usually not enabled, even in "debug" builds, since there was no
automatic mechanism to do so -- enabling them required manual changes
to the Kbuild file by individuals producing builds.  So we update the
Kbuild to automatically enable these option when certain debug options
are enabled in the Kernel config.

613450: Fix for low throughput in 11B mode to pass WFA test case 4.2.23
WMM parameters for 11B mode were not correct. So fix is it to correct
WMM params for 11B mode (CR 336266)

613617: Fix issue with purging PE Message Queue during shutdown
During Stress & Stability testing an issue was observed where, during
WLAN driver shutdown, a kernel BUG was logged in the kernel log with a
stack trace that indicated peFreeMsg() was invoking kfree() with an
invalid memory pointer.  Analysis of the code shows that a PE message
can contain either a pointer to allocated memory or a pointer to a
packet, but the logic within peFreeMsg() only correctly handles the
first case, and presumably the BUG occurred when peFreeMsg() invoked
kfree() with a packet pointer.  To fix the issue we incorporate logic
to handle both allocated memory and packet pointers. (CR 336061)

613683: Update cfg.ini options
Adding ini file option to disable /enable packet filtering. Also added
API to change the value of data inactivity timeout and enable/disable
scan (CR 334712)

613754: Fix WDI response dispatch table
Even though there is no shutdown response coming back from HAL, the
function is added there for the symmetry.  But this function is not
placed in the dispatch table, and because of this the indicators were
broken

Change-Id: Ib36ed26df3862fb2a72222b717507642ebcb4d23
CRs-Fixed: 334712,336061,336266