wlan: CS 1.0 preliminary release #29
Integrate the following CLs from Perforce:

614364: MAC Address auto-generation
Introduce functionality to auto-generate MAC addresses if the
addresses in NV are invalid and the addresses in cfg.ini have not been
modified from their default values.  This will reduce the likelihood
that the default addresses are used. (CR 334945)

614365: Enable MAC Address auto-generation

614372: Remove hard tabs introduced by changelist 614178

614403: Remove BSS when removing SoftAP
When hostapd get killed, the function wlan_hdd_cfg80211_del_beacon is
not getting called to remove the bss. So next time if we start the bss
without unloading the driver, it is resulting in failure to start bss.
Changes done to call HDD functions to delete bss and to stop hdd
adapter before deleting softap hdd adapter.

614409: Fix Memory leak in wlan_qct_wda.c
Memory allocated for baActivityInd message in WDA should be freed in
LIM. If the session for the BSSID in this message is not found then
the message is not getting freed. This result in memory leak. (CR
336342)

614410: Changes for setting the power params dynamically
1) support for setting power params from iwpriv.
2) Added the support for gEnableDynamicDTIM in cfg.ini. If this value
is 0, there is no change in DTIM used by SLM in suspend/resume
state. If this values is 1-5, SLM will switch to DTIM count specified
here when host goes into suspend mode and switch DTIM1 when host
resumes.

614471: Fix issue with DRIVER SET POWERMODE command
In BMPS disabled case, our driver is returning "FAIL" for setPowerMode
which is resulting in multiple load\unload of wifi driver (CR 337181)

614472: Changes for not dropping the beacons if we are in the beacon
missed scenario, to avoid hitting the HB failure.

614492: Fix TKIP+WEP keying issue
When AP is confiugred is in TKIP+WEP encryption mode, then supplicant
on STA configuring TKIP key as unicast key and WEP key as broadcast
key. But the driver is configuring both keys as unicast keys
only. Issue: nl80211 interface is passing a argument (pairwise) which
tell whether the key is pairwise or group, but our driver is not using
this arugment.

614504: 11D CH13 Scan issue fix
If NV.bin has US CC, first passive 11D scan does not scan CH12, 13,
14.  After regulatory domain chaged, during long time, old SCAN list
not in CH list still seen on AP list (CRs 337423, 333470)

614558: BT-AMP cleanup
- KW fix
- PAL qualification fix, Phy link validity check in read local amp assoc
(CRs 338182, 337889)

614562: Add CFG item to support LTE COEX

614601: Reverting the change made for CR 337889

614606: Remove wakelock(suspend lock) in WLAN driver
Remove wakelock added in WLAN driver; instead of using a wakelock,
fail the driver 'suspend' if Riva is not put in to IMPS/BMPS states.
This new method will also allow APPS to power collapse in SoftAP mode
(CR 337690)

614645: Allow enabling/disabling the APC and 40MHz SGI from cfg.ini
Fix for CR Need ini parameter to disable/enable APC & SGI (CR 338004)

614652: Properly clean up when concurrent SoftAp interface is removed
When we kill hostapd, the softap interface is not getting cleaned
properly before removing the interface.  If we use "kill <pid>" to
stop the hostapd, it is calling wlan_hdd_cfg80211_beacon first then it
is calling wlan_hdd_cfg80211_change_iface.  In
wlan_hdd_cfg80211_change_iface function driver is uniinitializing
softap adapter and changing interface type to station and
reinitializing as STA adapter.  Once we execute "iwpriv wlan0 exitAP"
it is not unitializing STA adapter before removing the interface.  If
we use "kill -9 <pid>" to stop the hostapd, it is not calling
wlan_hdd_cfg80211_beacon to delete the bss. This case was already
taken care of in a previous check-in (CR 338529)

Change-Id: I0e8f60ee7e417fb237a02a1f62488ae8d2c19409
CRs-Fixed: 333470,334457,334945,336342,337181,337423,337690,338004,338182,338529