wlan: CS 1.0 preliminary release #30
Integrate the following CLs from Perforce:

614821: Add support for Get Channel List (CR 335764)

615025: Fix TRP/TIS failure with Rhode&Shubaltz test equipment
Association for TRP/TIS fails with Rhode&Shubaltz test
equipment. TRP/TIS worked with Anritsu device. This issue seems to be
the same issue as CR 319225 (for WCN1314). The Rhode equipment send
out association response with 15 basic rate set as the previously
filed CR for WCN1314 (CR 330470)

615116: Support Display MAX data rate instead of current data rate
Some customer want to display supported MAX data rate instead of
current data rate. This feature should be configurable, since we
should not propagate this change to all customers (CR 331816)

615167: Adding New API to dynamically change mode of operation for ACL list
iwpriv setAclMode n where n can have value from 0 to 3 (CR 338666)

615205: LTE-WLAN COEX REQUIREMENT FOR SAP
When gEnableLTECoex is set to 1 in the INI configuration file, the SAP
would not select channels 10-14 during its initialization in
sapGetChannelList () as those channels are considered to be affected
by LTE interference. For this feature to take effect,
gApAutoChannelSelection in the INI configuration file need to be set
to 1 as well.

615210: Remove hard tabs introduced by CL 615205

615277: Fix issue with updating dynamic DTIM when driver is not in BMPS
Driver may not be in BMPS when early suspend kicks in, so only update
the config in this case & wait for the actual 'enter BMPS' request to
update the Dynamic DTIM (CR 339147)

615306: Remove hard tabs introduced by CL 615025

Change-Id: I3efd907b8afa0e5560687d132ef2998701667759
CRs-Fixed: 330470,331816,335764,338666,