wlan: Fix watchdog bite due to unclocked register access
During WLAN suspend/resume and stability testing multiple occurrences
of system restarts due to watchdog bites were observed.  Extensive
debugging revealed cases where the WLAN Host Driver was accessing RIVA
registers when they were in power collapse.  The code is modified to
fix those instances so that register accesses only take place when the
registers are powered on.  This is an integration of changelist
615910 from Perforce.

Change-Id: Iac4cf6ec6450303fa54e3c98734dc7e42d4d56b5
CRs-Fixed: 329903