wlan: CS 1.1 release #7
Integrate the following CLs from Perforce:

621373: Fix issue where SAP uses the wrong SSID
There was a bug in beacon generation logic where it would rely on cfg
variable to update the SSID.  If SAP comes up and it uses the SSID set
for global CFG. Later if STA comes up it will reset the CFG
variable. Now when SAP continues it will start using the STA's SSID
from CFG variable.  The issue is fixed by having SAP use the session's
SSID instead of the global CFG SSID.

621510: WAPI Rekey fix for prima
TL was not receiving the WAPI auth packets with the expected
RXBD->dpufeedback header values. So, TL code is modified to skip the
verification related to the check of 'dpufeedback' in order to
successfully complete the rekey negotiation. (CR 340139)

621683: SCAN request race condition fix (CR 345354)

621740: Fix IOT issue with ATT Uverse AP
ATT Uverse AP sending the probe repsonse with space as SSID for a
broadcast SSID proberequest when it is configured as hidden SSID. This
result in the scanresult with actual SSID is overwritten with the SSID
as "space". To fix this added the logic to consider the SSID as NULL
if all characters in SSID are spaces. (CR 349166)

621747: Fix Memory leaking in limCheckAndAddBssDescription()
Not freeing the BssDescription if the beacon/proberesponse recieved
with NULL SSID and there is already an entry in the
limcachetable. Added to code to free this memory. (CR 346945)

621828: Fix kernel panic in vsnprintf()
A kernel panic was observed in vsnprintf() when invoked from the WLAN
logDebug() API.  Analysis of the logDebug() API showed that the
implementation was not reentrant, and the resulting buffer could cause
random memory corruption.  The implementation is changed to be
reentrant and to not use the generated log as an output format
string. (CR 344945)

621931: Add copyright information to prima files (CR 330775)

621945: Correct line continuation in debug message (CR 348777)

621948: Fix rate set validation
Need to check whether extended supported rates too along with
supported rates are present or not

622086: Fix Memory corruption when unloading prima_wlan
In customer trials it was observed that on occasion during wlan driver
unload that the message "VPKT [113]: Invalid magic" was observed.  The
root cause is that the Message Buffer free pool becomes exhausted, and
the DXE driver is not able to be successfully closed.  As a result the
Packet Buffers become corrupted.  To fix this we increase the number
of Message Buffers.  In addition we update the Packet Buffer code to
"give up" when it encounters an invalid magic while destroying a
packet list. (CR 343699)

622191: Update copyrights for newly modified files

622353: Fix infinite loop during vos_open()
The recent change to increase the number of VOSS Message Buffers had
an unexpected side effect.  vos_open() was using an U8 to iterate when
it initialized the message free pool, but since the number of buffers
now exceeds 255, the loop would never exit.  Fix by using a U32 for
the iterator.

Change-Id: I3b620fc2c2d4b967630516c9a42b2e0b2382f14f
CRs-Fixed: 330775,340139,343699,344945,345354,346945,348777,349166,