wlan: VOWIFI defeaturization and supporting backward compatibility
Restore most of the functionality of the feature "VOWIFI
defeaturization and supporting backward compatibility" that was
originally checked in to Perforce but which was reverted since the
feature did not work with deployed Riva 1.0.  We now support reporting
the host version of the Riva API, but we do not include the messages
which were initially modified.

Change-Id: If03d0f831b3ac34b19f2797b23d9255cc491983c