wlan: Prevent kernel panic in cfg80211_rx_mgmt()
A kernel panic was observed in cfg80211_rx_mgmt() when invoked from
hdd_indicateMgmtFrame().  Analysis of the stack trace indicates that
cfg80211_rx_mgmt() was invoked with an invalid device pointer.  In
order to prevent this in the future we add "magic" to the Adapter
structure so that we can verify that the Adapter is in a consistent
state, and we add consistency checks to hdd_indicateMgmtFrame().

Change-Id: I720dbfbd4808e80bbb83e707e86465c491e7d33b
CRs-Fixed: 349016