wlan: Cleanup VOSS ERROR logging
Previously when we enabled VOSS logging at the ERROR level we would
get some spurious messages which were not indicators of errors.  In
addition we would get some messages which would seem to indicate
errors.  We resolve these issues by examining the source of the
messages and correcting either the logic or the logging level as
appropriate.

Change-Id: I372a740a9a84ea82c82c8d40228a58d10a03e30d
CRs-Fixed: 351625