wlan: Fix for mobile hotspot unable to turn off
The function sme_ProcessMsg acquires the sme_AcquireGlobalLock see the
Call Stack 1 below. In the meantime if the wifi is turned off the
kernel calls the function cfg80211_change_iface and acquires the
RTNL_lock and the function sme_AbortMacScan blocked on
sme_AcquireGlobalLock because the sme_ProcessMsg already taken that
lock see the Call Stack 2.

Now sme_ProcessMsg calls the hdd_hostapd_SAPEventCB which calls the
flush_workqueue blocked on RTNL_lock because the cfg80211_change_iface
taken the RTNL_lock.  sme_ProcessMsg release the Sme global lock only
the hdd_hostapd_SAPEventCB function returns but this function is
blocked because of the RTNL_lock causing the dead lock.

Call Stack 1
-------------
[<c018e67c>] flush_workqueue+0x154/0x334
[<bf0e4fcc>] hdd_hostapd_SAPEventCB+0x31c/0xb98 [wlan]
[<bf080820>] sapSignalHDDevent+0x464/0x4dc [wlan]
[<bf03f79c>] csrProcessDelStaSessionRsp+0x7c/0x100 [wlan]
[<bf05bb5c>] sme_ProcessMsg+0x2e4/0x550 [wlan

Call Stack 2
-------------
[<bf0016f8>] vos_lock_acquire+0xb4/0x104 [wlan]
[<bf057ff4>] sme_AcquireGlobalLock+0x14/0x24 [wlan]
[<bf0594b8>] sme_AbortMacScan+0x18/0x3c [wlan]
[<bf0c7e34>] hdd_stop_adapter+0xe8/0x290 [wlan]
[<bf0edebc>] wlan_hdd_cfg80211_change_iface+0x23c/0x378 [wlan]
[<c0785f80>] cfg80211_change_iface+0x100/0x1c0
[<c079116c>] nl80211_set_interface+0x188/0x1c4
[<c066ff10>] genl_rcv_msg+0x1d8/0x224
[<c066e338>] netlink_rcv_skb+0x4c/0xb0
[<c066fd2c>] genl_rcv+0x18/0x24

To fix this before calling the HDD callback release the SME global
lock and acquire the SME global lock after the HDD callback returns

Change-Id: I4a06e7645735e8dad0ae9c61ad3359b1a888830b
CRs-Fixed: 347642, 349420