wlan: CS 1.1 release #8
The following CLs have been integrated since the last release:

618627: VOWIFI defeaturization and supporting backward compatibility
Restore most of the functionality of the feature "VOWIFI
defeaturization and supporting backward compatibility" that was
originally checked in to Perforce but which was reverted since the
feature did not work with deployed Riva 1.0.  We now support reporting
the host version of the Riva API, but we do not include the messages
which were initially modified.

622203: Fix race condition during SSR
Some of the wext interfaces are not SSR protected, adding protection
to these APIs. (CR 345055)

622403: Proposed fix for management path timeout/crash
When WDA times out waiting for the TX path to send out a management
frame it informs PE about the failure; PE destroys the frame in the
MC thread context.  In the mean while the frame can be anywhere in the
TX data path components and its deletion will cause an unexpected
behavior that will more often then not lead to a crash.  The proposed
fix prevents WDA from notifying failure to PE, instead it tags the
frame as timed out using the WDA user data id from the vos packet .
When frame completes tx it will end up in WDA marked appropriately -
WDA will just discard the packet at its level and it will not do
anymore processing as this isn't the current frame that it is trying
to send out.  Unblocking the MC thread before a frame has completed tx
opens up the possibility of UMAC sending another mgmt frame to TL
while the 1st one is still pending TX at TL level. In this case to
eliminate any possible race condition and not to introduce locks TL
will just fail the TX request.  In all customer reports the crash
seems to happen when the frame was already on its completion path and
the time gap between the timeout and crash is very small, reason for
which I also increased the WDA timer value to minimize the likelyhood
of this occurance. (CR 339973)

622495: Handle back to back Remain On Channel Request
If a new Remain On Channel Request is received while a Remain On
Channel Request is pending then allow the new one by cancelling the
old one. (CR 346946)

622517: Fix scan result getting purged while doing first scan
After receiving first scan result, while doing change domain we are
calling csrScanFlushResult function which will remove all scan result
entry from scan list. Due to this we always get zero result for first
time scan which is creating long time to display the scan list in
GUI. Now instead of flushing all scan result entries we will flush
only those scan results which correspond to channels which are not in
the new channel list. (CR 348076)

622528: Handle WDI timeout during WLAN driver unload process
If a WDI timeout happens while unloading the driver, complete the
unload using the 'shutdown' method & then at the end of module exit
invoke SSR. (CR 343719)

622598: Keep 11h setting in sync with 11d
Enabling 11h based on 11d, If 11d is enabled then we need enable 11h
too else disabled. (CR 344846)

622743: Scan command pending list, extra protection (CR 348034)

623034: Fix improper beaconing of country IE by SAP
SAP is advertising country information element in Beacon even if
802.11d parameter is disabled.  ieee80211d parameter configuration in
hostapd.conf is not propagated to LIM; only if Country IE is present
then country code is sent to LIM. LIM reads default configuration of
ieee80211d configuration which is only updated by Is11dSupportEnabled
on HDD startup and adds country IE based on that, so even if
hostapd.conf disables the country code, the setting is ignored. So
added new member in tCsrRoamProfile for ieee80211d configuration which
is used to update WNI_CFG_11D_ENABLED config, which later is used by
LIM while deciding whether or not to add the country code.  (CR
349577)

623114: Prevent kernel panic in cfg80211_rx_mgmt()
A kernel panic was observed in cfg80211_rx_mgmt() when invoked from
hdd_indicateMgmtFrame().  Analysis of the stack trace indicates that
cfg80211_rx_mgmt() was invoked with an invalid device pointer.  In
order to prevent this in the future we add "magic" to the Adapter
structure so that we can verify that the Adapter is in a consistent
state, and we add consistency checks to hdd_indicateMgmtFrame().  (CR
349016)

623145: Cleanup VOSS ERROR logging
Previously when we enabled VOSS logging at the ERROR level we would
get some spurious messages which were not indicators of errors.  In
addition we would get some messages which would seem to indicate
errors.  We resolve these issues by examining the source of the
messages and correcting either the logic or the logging level as
appropriate. (CR 351625)

623297: Fix for mobile hotspot unable to turn off
The function sme_ProcessMsg acquires the sme_AcquireGlobalLock see the
Call Stack 1 below. In the meantime if the wifi is turned off the
kernel calls the function cfg80211_change_iface and acquires the
RTNL_lock and the function sme_AbortMacScan blocked on
sme_AcquireGlobalLock because the sme_ProcessMsg already taken that
lock see the Call Stack 2.

Now sme_ProcessMsg calls the hdd_hostapd_SAPEventCB which calls the
flush_workqueue blocked on RTNL_lock because the cfg80211_change_iface
taken the RTNL_lock.  sme_ProcessMsg release the Sme global lock only
the hdd_hostapd_SAPEventCB function returns but this function is
blocked because of the RTNL_lock causing the dead lock.

Call Stack 1
-------------
[<c018e67c>] flush_workqueue+0x154/0x334
[<bf0e4fcc>] hdd_hostapd_SAPEventCB+0x31c/0xb98 [wlan]
[<bf080820>] sapSignalHDDevent+0x464/0x4dc [wlan]
[<bf03f79c>] csrProcessDelStaSessionRsp+0x7c/0x100 [wlan]
[<bf05bb5c>] sme_ProcessMsg+0x2e4/0x550 [wlan

Call Stack 2
-------------
[<bf0016f8>] vos_lock_acquire+0xb4/0x104 [wlan]
[<bf057ff4>] sme_AcquireGlobalLock+0x14/0x24 [wlan]
[<bf0594b8>] sme_AbortMacScan+0x18/0x3c [wlan]
[<bf0c7e34>] hdd_stop_adapter+0xe8/0x290 [wlan]
[<bf0edebc>] wlan_hdd_cfg80211_change_iface+0x23c/0x378 [wlan]
[<c0785f80>] cfg80211_change_iface+0x100/0x1c0
[<c079116c>] nl80211_set_interface+0x188/0x1c4
[<c066ff10>] genl_rcv_msg+0x1d8/0x224
[<c066e338>] netlink_rcv_skb+0x4c/0xb0
[<c066fd2c>] genl_rcv+0x18/0x24

To fix this before calling the HDD callback release the SME global
lock and acquire the SME global lock after the HDD callback returns
(CRs 347642, 349420)

623320: DEL STA after STA's TX RING empty (CR 350079)

623345 & 623357: Set Listen Interval in HDD context
Keep track of the Listen Interval in HDD.  Otherwise under certain
race conditions HDD will set the Listen Interval to 0 and RIVA will
crash. (CR 351810)

Change-Id: I5b2137c5d2357027e8e39231933a58f903fd2e14